7-15
연구탭의 건축, 원자, 메탈, 크리스탈, 가스 정제는 연구레벨에 따른 상승폭 적용됨.

에너지 공학 연구 레벨에 따른 전력 생산량 증가 및 그 외 로직들 적용해야함.
너무 중구난방적으로 흩어져 작성되어 있음으로 연구 및 특수 능력 적용 로직은 PlayerAbilityInfo의 스크립트를 활용하자

+++ Assets/Script/Ui/Scriptable/3.연구/Research.cs
@@ -26,7 +26,7 @@ public class Research : ScriptableObject
     public int[] research_Time; // 업그레이드 시간
     public int[] research_Cost; // 업그레이드 비용 (메탈/크리스탈/가스)
     public float[] upgrade_Cost_Require; // 레벨당 업그레이드 비용 증가 배수
-    public float research_Ability; // 연구 레벨당 증가량
+    public int research_Ability; // 연구 레벨당 증가량
     public string ability_Text; // 연구 내용
     public int[] miningSlot; // 채광 레벨당 함대 슬롯
